tests green, remove old schema
