+++ lib/graphql/api.rb
@@ -1,22 +1,26 @@
 require "graphql/api/version"
 require "graphql/api/policy"
-require "graphql/api/schema"
+require "graphql/api/command_type"
+require "graphql/api/query_type"
 require "graphql/api/configure"
 
 module GraphQL
   module Api
 
     def self.schema(opts={})
-      GraphQL::Api::Schema.new(opts).schema
-    end
-
-    def self.graph(opts={})
-      GraphQL::Api::Schema.new(opts)
+      @schema ||= configure_default_schema(opts)
     end
 
     def self.configure(&block)
       config = GraphQL::Api::Configure.new
       config.instance_eval(&block)
+      @schema = config.schema
+      @schema
+    end
+
+    def self.configure_default_schema(opts)
+      config = GraphQL::Api::Configure.new
+      config.with_defaults(opts)
       config.schema
     end
