--- lib/graphql/api/schema_error.rb
@@ -1,4 +1,0 @@
-module GraphQL::Api
-  class SchemaError < Exception
-  end
-end
