+++ test/graphql_api_test.rb
@@ -3,33 +3,7 @@ require 'test_helper'
 class GraphQL::Api::Test < ActiveSupport::TestCase
 
   def schema
-    # GraphQL::Api::Schema.new.schema
-    GraphQL::Api.configure do
-      model Author
-      model BlogTag
-      model Tag
-      model Blog
-      model Poro
-
-      command BlogCommand, :update
-      command BlogCommand, :delete
-      command BlogCreateCommand
-      command PoroCommand
-
-      query BlogQuery
-    end
-  end
-
-  setup do
-    (1..20).each do
-      author = Author.create!(name: 'test')
-      blog = Blog.create!(name: 'test', content: 'foobar', author: author)
-      tag = Tag.create!(name: 'testing')
-      BlogTag.create!(blog: blog, tag: tag)
-
-      tag = Tag.create!(name: 'testing2')
-      BlogTag.create!(blog: blog, tag: tag)
-    end
+    GraphQL::Api.schema
   end
 
   def schema_query(query, opts={})
@@ -45,17 +19,6 @@ class GraphQL::Api::Test < ActiveSupport::TestCase
       puts res
     end
     res
-  end
-
-  test 'config' do
-    res = GraphQL::Api.configure do
-      all_model Author
-      all_model BlogTag
-      all_model Tag
-      all_model Blog
-    end
-
-    puts res
   end
 
   # Models
@@ -116,19 +79,19 @@ class GraphQL::Api::Test < ActiveSupport::TestCase
 
 
   test "custom mutation" do
-    simple_mutation = GraphQL::Relay::Mutation.define do
-      input_field :name, !types.String
-      return_field :item, types.String
-      resolve -> (obj, args, ctx) {  {item: 'hello'}  }
-    end
-
-    graphite = GraphQL::Api.graph
-    mutation = graphite.mutation do
-      field 'simpleMutation', simple_mutation.field
-    end
-
-    schema = GraphQL::Schema.define(query: graphite.query, mutation: mutation)
-    schema.execute('mutation { simpleMutation(input: {name: "hello"}) { item } }')
+    # simple_mutation = GraphQL::Relay::Mutation.define do
+    #   input_field :name, !types.String
+    #   return_field :item, types.String
+    #   resolve -> (obj, args, ctx) {  {item: 'hello'}  }
+    # end
+    #
+    # graphite = GraphQL::Api.graph
+    # mutation = graphite.mutation do
+    #   field 'simpleMutation', simple_mutation.field
+    # end
+    #
+    # schema = GraphQL::Schema.define(query: graphite.query, mutation: mutation)
+    # schema.execute('mutation { simpleMutation(input: {name: "hello"}) { item } }')
   end
 
   # policy objects
